+++ crates/nord-format/src/formats/raw.rs
@@ -25,7 +25,7 @@ macro_rules! raw_format {
                 /// The body length every corpus specimen holds. Observed, never
                 /// enforced on read: a raw body cannot misread, so a file of
                 /// another length is preserved rather than refused.
-                pub const BODY_LEN: u64 = $body_len;
+                pub const BODY_LEN: usize = $body_len;
             )?
 
             pub fn read_from(reader: &mut (impl Read + Seek)) -> Result<Cbin<RawBody>, Error> {
